fix(auto-optimizer): cast to f64 before dividing memory autoscan progress
The memory OC phase computed `(usize + usize) / (usize + usize)` then
formatted it with `{:.2}%`, so the result was always truncated to 0 or
1 before the format specifier ran — printing `0.00%` for the entire
scan and `1.00%` briefly at the end.

Cast both sides to f64 and multiply by 100.0, matching the identical
pattern already used correctly in the core-OC short (L594) and long
(L874) phases.

Fixes #17

+++ auto-optimizer/src/oc_scanner.rs
@@ -1103,7 +1103,9 @@ fn run_mem_oc_phase<V: std::fmt::Display + Copy>(
 
         println!(
             "current test progress estimated:{:.2}%",
-            (mem_test_num + mem_test_code - 1) / (args.mem_freq_step_exp + mem_test_code - 1)
+            (mem_test_num + mem_test_code - 1) as f64
+                / (args.mem_freq_step_exp + mem_test_code - 1) as f64
+                * 100.0
         );
         println!("current test num: {}", mem_test_num);
